Removing extra return from code
The code was returning if at least one profiling could not have its data fetched, stopping the whole processing. Now we just make it skip the profiling.

This will make the profiling upload never be considered again. But this is usually fine as this part of the code is not about the individual uploads, but about the statistical volume