Adding unexpected changes on new and deleted files properly
As described on https://github.com/codecov/worker/pull/746 , there are a couple of conditions where the unexpected changes might show/not show up on a file that is new on the report (present on head and not on base)

Namely, if a file is new on the report, and the diff says it's not a new file, we are currently fully considering it unexpected. But if the file is part of the git diff, and all the new lines on the report were lines added by the git diff, then it makes sense that this is an *expected* change.

To calculate that, we unfortunately need to calculate the `unexpected_vec` vector on even more situations now. This might produce a small performance impact (although we don't expect it to be big).