Limiting unexpected changes calculations
Unfortunately, we found that Github in some cases won't return the actual diff.

In one example, on a new file, here is what Github returned (names changed to hide customer actual name)

```
{'sha': 'dd770e6c90c8c6180e6196df587ebca8568274fc', 'filename': 'path/to/file/rewards/coin-drop.svg', 'status': 'added', 'additions': 9, 'deletions': 0, 'changes': 9, 'blob_url': 'https://github.com/owner/repo/blob/a22a46bbdf1d2b7ddd474be053da602edbeab95d/path/to/file/rewards/coin-drop.svg', 'raw_url': 'https://github.com/owner/repo/raw/a22a46bbdf1d2b7ddd474be053da602edbeab95d/path/to/file/rewards/coin-drop.svg', 'contents_url': 'https://api.github.com/repos/owner/repo/contents/path/to/file/rewards/coin-drop.svg?ref=a22a46bbdf1d2b7ddd474be053da602edbeab95d'}
```

This case happened on an "added file". It is possible that Github will require an extra get to the `raw_url` field or something to get the actual diff.

So my guess is that this happens just like in the UI: on added and deleted files (that might be too big?)

Unfortunately, on the rest of the files we have no way of fixing this except for actually making those extra api calls. But on new and deleted files, we know that there won't be any unexpected changes (unless a git-new file somehow was present on the base report, but that only happens on a bug either on our end or on the customer lib)

An alternative to this is: on a git-new file we can know, from looking which lines the new report has, at least a few of the lines that were added. For example, if a diff says a certain file is new and its reportfile has line 103, we can consider that the git diff added at least all the lines from 1 to 103. But I think that brings more complications than our chosen solution