Removing asyncio.run from code
In previous changes, we found that asyncio.run is causing some pages hanging.

We recently found a case where this is happening. And while we can't find which place the asyncio.run could be causing it to hang, at this point it is probably easier to remove a few more asyncio.run calls and see if it improves the situation

There are still a couple of asyncio.runs left. But those are the ones that touch UI stuff