Using codecov opentelem on API
Implementing the call to the `get_codecov_opentelemetry_instances` on the opentelemetry point

This is done on the `wsgi`, because this seems that this is the entrypoint for the web application (and this is where this was done before on the research phase). Happy to change to a different place if that helps.

Just like in the worker, this will require `OPENTELEMETRY_ENDPOINT`, `OPENTELEMETRY_TOKEN` and `OPENTELEMETRY_CODECOV_RATE` which all affect how the instrumentation will work.

Also adding a `get_current_version` function to the APi. We have yet to implement the envvar on the dockerfile, but at least this unblocks us on that front for now with a hardcoded version