Creating new endpoint for profiling
Creating new endpoint for profiling:

```
    path(
        "versions", ProfilingCommitCreateView.as_view(), name="create_profiling_version"
    )
```

This is a path where customer can create a "profiling version". The idea here is that, with an `environment` field being added, the `version_identifier` stops being an unique field (in the hopes that this will allow the customer to upload two different "datasets" to the same version: production and end-to-end-testing. So they would have the same `version_identifier` but different environments.

So what we are doing here is:
- Creating an endpoint where the customer can POST to create a (or update an existing one) new "profiling version"
- Changing the other endpoint to not use the `version_identifier` anymore, and instead use an `external_id`, an unique field
- The opentelem libraries will at startup call the new endpoint to create a new version. The API will respond with its `external_id`.
- The client from then on will use that `external_id` to send all the spans to the already-existing endpoint

This change will require on more API call at the startup of the servers, which customers might not like. We made it so they can make the call to that endpoint elsewhere and directly embed the `external_id` as a parameter to the setup call. That does make it so they have to make the call from somewhere and pass to their system as an env-var or something. Which I don't know whether is a problem for them or not

An alternative is to create a truly unique `code` column on that model, and let users just use that instead of getting an `external_id` dynamically. I have yet to determine whether that is needed or not.