Renaming Upload-related models
For some reason the models on API follow the `ReportSession*` naming structure. They should follow the `Upload`, like the worker does.

Unfortunately, to properly rename them, we need to edit the migration a bit in order to not let it drop and recreate a contraint with another name, so for now the `ReportSession` table will remain as it is.

But we will do this change soon