Throttling uploads per user
The idea here is relatively straightforward: if a given user has an "upload-restricted" plan, we want to enforce that restriction at upload time and only on the first upload a certain commit is given.

And we want to create a "users-basic" plan for the free customers and add a 250 uploads per month restriction to it.

This is what this change is doing with a slight modification: we are only checking uploads make in the past month and only for commits that have timestamps for the past two months. This is to tell the db to not attempt to look at very old commits to find the uploads, because while technically it could be possible, it's unlikely in reality that very old commits receive new uploads (although there are certainly cases), and this makes the query way slower (from 0.3s to 4s in an anecdotal case). We hope that as we see this query in practice, we can create a better index to power it, and this problem can go away and we can remove this limitation.

The database query generated by it is:

```
SELECT COUNT(*)
FROM
  (SELECT "reports_upload"."id" AS Col1
   FROM "reports_upload"
   INNER JOIN "reports_commitreport" ON ("reports_upload"."report_id" = "reports_commitreport"."id")
   INNER JOIN "commits" ON ("reports_commitreport"."commit_id" = "commits"."id")
   INNER JOIN "repos" ON ("commits"."repoid" = "repos"."repoid")
   WHERE ("reports_upload"."created_at" >= '2021-09-06T22:32:59.231162+00:00'::timestamptz
          AND "repos"."ownerid" = 2
          AND "commits"."timestamp" >= '2021-08-07T22:32:59.231184+00:00'::timestamptz
          AND "reports_upload"."upload_type" = 'uploaded') LIMIT 250) subquery
```

which ensures the count only goes up to 250, so it doesn't take too long

On the next few changes we will create indexes to optimize this query. But as it stands (specially that this will only apply to new users), this should be fine.

It is also only generate if the user is in one of the limited plans, so we can reduce impact for paying customers (that are not on monthly limited plans)