Making get_changes account for removals
There was one case where the python get_changes code was not accounting for:

If a file was present on base, not on head, but was enumerated on the diff files, the code would simply consider it accounted for. This works in some cases like when a file was fully deleted.

But this didn't consider that the diff might not have accounted for the full deletion of the file. For example, if a file had lines 10, 11 and 12 and the diff removed lines 11 and 12, we should still expect line 10 to be there. In this case, if the whole file is missing on head, then the removal of line 10 was unexpected