Let Dev API Parse codecov.yml config
In the course of doing local development, it may be necessary to allow requests from more sources than just `localhost`. This commit allows the config yaml to be parsed for `setup.api_allowed_hosts` similar to the enterprise configuration. This still defaults to `localhost`, so it should work as intended for developers if they do not have `setup.api_allowed_hosts` provided in their configurations.