Creating well defined enums for upload state and upload type
As we move to rely more on the `Upload` table from the database, we need to make more queries on it.

And there are two columns on that table, `state` and `upload_type`, which have `Text` types despite having a limited number of options they could be (specially because those are all values defined by the system and not by the user).

So it makes sense to move them columns torwards an integer type. This will make the table a bit safer (no room for arbitrary strings), and smaller for creating indexes on those columns.

To do that, we are creating the resposible enums here in `shared`. This will allow API and worker to use consistent values