Temporarily hiding some unexpected changes cases from rust changes
On python, in order to determine the unexpected changes between two versions of `ReportFile`s, we usually iterate them line by line comparing the base line to the head line. A line is considered unexpectedly changed if its type changes (hit -> partial, miss -> nonexistent, partial -> miss etc).

But in some cases, like partials, two lines can be both partials and still have different results. For example, in one partial we could have a 5/6 coverage (meaning 5 of the 6 possible "sections" of the partial are covered) and on other a 3/6. This means that the line coverage changed unexpectedly.

On python, that is not treated like a change (neither expected or unexpected). In rust, we do treat those two line as unexpectedly changed, leading to more unexpected changes cases than in python.

While the rust behavior seems more correct (something in the system did affect the coverage of a section of the code), our UI and system don't tend to tell the customer about this level of change. So it might be complicated to say the line unexpectedly changed and then, when the customer goes to our UI to check, show everything being the same.

So we are doing a "band-aid" change while we decide the way to go.

This "both partial but different values" end up bubbling up to the `_get_changes_from_comparison` function like:

```
                    "unexpected_line_changes": [
                        [[1, "p"], [1, "p"]],
                        [[2, "p"], [2, "p"]],
                    ],
```

so in the code we are just filtering out this case for now.

If we decide to keep python way, we implement a proper check in the rust code. If rust is, we will make an attempt to properly bubble up this data to the python level on this function.