Creating `code` field on CommitVersion
As described on https://github.com/codecov/codecov-api/pull/815 :

> An alternative is to create a truly unique code column on that model, and let users just use that instead of getting an external_id dynamically. I have yet to determine whether that is needed or not.

It was decided that on PHP, the frameworks designs make it prohibitively expensive to make one API call to exchange the variables for an `external_id`, since there is no "application scope", and everything lives and dies when a request is received/responded.

So we are adding such a column, with the following additions:

- The `code` starts being required on the `POST /profiling/versions` call
- There is a unique constraint on `(repoid, code)`, ensuring that customers can use each code only once per repo
-  The `POST /profiling/uploads` will start using `code` instead of `external_id` (keeping both would make the API very confusing)

There will also be a migration involved in this change. Here its SQL:

```
BEGIN;
--
-- Add field code to profilingcommit
--
ALTER TABLE "profiling_profilingcommit" ADD COLUMN "code" text NULL;
--
-- Create constraint uniquerepocode on model profilingcommit
--
ALTER TABLE "profiling_profilingcommit" ADD CONSTRAINT "uniquerepocode" UNIQUE ("repoid", "code");
COMMIT;
```

The constraint will require an index to be created. But since:

1. This table is not used by anyone but us for now
2. Even we are not using this table for now
3. The table has exactly 12 rows

It should not be a problem that this table will be locked during the migration.