Changing serializer update
As described on https://github.com/encode/django-rest-framework/issues/2648 , DRF won't implement a way of doing a partial update.

Unfortunately, due to the constant flow of API calls on profiling, sometimes when a POST is made to the `/profiling/versions` endpoint, it coincides with the time a "collection" or "summarization" is running. And on that case if Django does a normal save, it ends up overwriting the `last_joined_uploads_at`, `last_summarized_at`, `joined_location`, `summarized_location` fields.

While this is not really damaging if it happens only once (the next task will fix it), we don't know how common it will be in real life (maybe it happens often enough to delay the collection to work for many many hours).

We inspired on a solution from https://github.com/encode/django-rest-framework/issues/5897 and wrote a few tests on it. It seems to work decently.

Also using `get_or_create` in the view, since it captures more cases than our own-rolled exception catching