Creating Account table
Currently, the Owner table is overloaded with meanings. It corresponds to:

1. The entity that owns repositories (as in `codecov` owns `codecov/worker` repository
2. The entity that makes commits (as in `ThiagoCodecov` that made a commit to a certain repo)
3. The entity that logs in (whenever we click the login button)
4. The entity that has a paying account at codecov

So now we want to create the Account model. The idea is that Account holds the "paying account" (item 4) level stuff. It removes from the Owner the responsability of having the "paying account" data.

This for examples solves a good part of the Gitlab-subgroup problem. Now all subgroups will belong to the same Account entity as the Group. All billing information will be centralized in one entity, and changing something on the top-level group won't require us to go manually on each subgroup and doing the same changes. And eventually (see below why), we will be able to activate users only once on that Account, and they will be "automatically" activated on every subgroup.

The following fields that would be "moved" from Owner (for a while we won't remove them from the table)

- stripe_customer_id
- stripe_subscription_id
- plan
- plan_provider
- plan_user_count
- plan_auto_activate

More fields might need to move eventually to Account, but those are the ones that we know are relevant in an Account context.

For example, we will eventually need to move `plan_activated_users` to a many-2-many between Account and Owner. This change will take a while to do, but will eventually help disassociate the "activated user" concept from being part of the Owner. This for example will allow us to more easily allow out-of-the-organization owners still be activated to use it.

This is the generated SQL migration

```
BEGIN;
--
-- Create model Account
--
CREATE TABLE "codecov_auth_account" ("id" bigserial NOT NULL PRIMARY KEY, "external_id" uuid NOT NULL, "created_at" timestamp with time zone NOT NULL, "updated_at" timestamp with time zone NOT NULL, "stripe_customer_id" text NULL, "stripe_subscription_id" text NULL, "plan" text NULL, "plan_provider" text NULL, "plan_user_count" smallint NULL, "plan_auto_activate" boolean NULL);
--
-- Alter field owner on ownerprofile
--
--
-- Alter field key on repositorytoken
--
--
-- Alter field valid_until on repositorytoken
--
COMMIT;
```