Fixing line removals
We have a bug on line removals where we are incorrectly calculating what the removed lines were. We were neither considering the segment starting difference between base and head and we were subtracting the wrong value.

Fixed those and added tests to ensure the tests work (checked manually that the results are correct)