Attempting image upgrade
Seeing if the system would work on:

```
python:3.9.7-alpine3.14
```

Unfortunately, there is this change on alpine 3.14: https://wiki.alpinelinux.org/wiki/Release_Notes_for_Alpine_3.14.0#faccessat2

It leads to problems like

```
make: mkdir: Operation not permitted
```

Locally, it works fine, but that page indicates that we need some things to make it work. One of the options is a more recent version of docker. So we set it to `20.10.7` which is the most recent one available on CircleCI.