Fixing eof usage on `iter_changed_lines`
In our system, on the `get_changes` calculations, we pretty much, for each file, align the base lines with their corresponding head lines (accounting for the git diff), and see if the coverage changed between both (miss becoming a hit, hit becoming a "nothing" etc).

On that pairing, we use the lines on the head file as reference, and from there we calculate what the corresponding line number for the base is.

We could potentially just say: "let's look at all the lines up to 10 million" and it wouldn't hurt us result-wise. We would just have a bunch of `None` that turned into `None`s, which have no bearing on the final result. We just don't do it because performance-wise that doesn't make sense

So we do some logic for "up to which head line number do we look?". Currently the logic is:

```
        for ln in range(
            1,
            max(
                (
                    base_report_file.eof if base_report_file is not None else 0,
                    head_report_file.eof,
                )
            )
```

which just picks the highest between base eof and head eof (eof in this context is simply the highest line that has some coverage information)

But this logic misses one case: when the user adds a ton of lines between head and base.

For example, let's say a file has only line number 2 covered on base and head. But they are adding 20 lines to the beginning of the file between base and head. We should expect to compare the line 2 on the base to line 22 on head, and point out there is a change there (because line 22 is missing on head). But since head head eof == 2, it doesn't do that comparison, because the system looks at max(2, 2), and determines it only needs to look at the lines up to 2 on head, making the (2, 22) comparison never happen.

This change solves that. It looks at the number of lines removed and added to the file on that diff, and then we tell the system "look at least until `base_file.eof + the delta of lines`". This way we guarantee that the lines are all analyzed.