Changing Bitbucket context sizes on diffs
Look at the following git diffs generated by *removing lines 15 and 16* from a given file (a simplified example). We will run the same command with just different context sizes. From 4 to 0.

```
➜  samplerepository git:(master) ✗ git diff -U4
diff --git a/README.md b/README.md
index 87f9baa..51c8a2d 100644
--- a/README.md
+++ b/README.md
@@ -11,10 +11,8 @@ ait suum! Ubi *insula*, phocae aerata, quas herba deficiunt, desiderioque ista
 imbres *et spectatam* cannas vellem!

 Truces luctuque cognovit, cum lanam ordine vereri relinquunt sit munere quidam.
 Solent **torvi clamare successit** ille memores rogum; serpens egi caelo,
-moventem gelido volucrum reddidit fatalia, *in*. Abdit instant et, et hostis
-amores, nec pater formosus mortis capiat eripui: ferarum extemplo. Inmeritas
 favilla qui Dauno portis Aello! Fluit inde magis vinci hastam amore, mihi fama
 et vulnus neque sanguine.

 Conversa domosque
```

If you look at the file and the numbers (-11,10 +11,8), it all makes sense. The "ait.." line is kind of a title. And then the "imbres..." line, which is the actual first line in the diff, is line 11 on that file. So both base and head correctly tell where this patch starts. Line 12 is the empty line, "Truces" is 13 and "Solent" is 14. Then the "-" signs show that line 15 (moventem) and 16 (amores) are being removed. Makes sense.

The documented pattern on unified diffs (which diffs implement) is **kinda** this: (start_line_base, base_patch_size, start_line_head, head_patch_size). Looking at the above, it makes sense. We can see it here: https://www.gnu.org/software/diffutils/manual/html_node/Detailed-Unified.html#Detailed%20Unified

Now with 3:

```
➜  samplerepository git:(master) ✗ git diff -U3
diff --git a/README.md b/README.md
index 87f9baa..51c8a2d 100644
--- a/README.md
+++ b/README.md
@@ -12,8 +12,6 @@ imbres *et spectatam* cannas vellem!

 Truces luctuque cognovit, cum lanam ordine vereri relinquunt sit munere quidam.
 Solent **torvi clamare successit** ille memores rogum; serpens egi caelo,
-moventem gelido volucrum reddidit fatalia, *in*. Abdit instant et, et hostis
-amores, nec pater formosus mortis capiat eripui: ferarum extemplo. Inmeritas
 favilla qui Dauno portis Aello! Fluit inde magis vinci hastam amore, mihi fama
 et vulnus neque sanguine.

```

Same thing: it all makes sense: (-12,8 +12,6). Line 12 is again the empty line. And counting from line 12, all the lines match the lines on the previous example.

So, what we see here (and what our code assumes on its implementations), is that 12 is the first line of this specific patch.

Now, with 2:

```
➜  samplerepository git:(master) ✗ git diff -U2
diff --git a/README.md b/README.md
index 87f9baa..51c8a2d 100644
--- a/README.md
+++ b/README.md
@@ -13,6 +13,4 @@ imbres *et spectatam* cannas vellem!
 Truces luctuque cognovit, cum lanam ordine vereri relinquunt sit munere quidam.
 Solent **torvi clamare successit** ille memores rogum; serpens egi caelo,
-moventem gelido volucrum reddidit fatalia, *in*. Abdit instant et, et hostis
-amores, nec pater formosus mortis capiat eripui: ferarum extemplo. Inmeritas
 favilla qui Dauno portis Aello! Fluit inde magis vinci hastam amore, mihi fama
 et vulnus neque sanguine.
 ```

 Notice that with context=4, the patch started at line 11. With context=3, the patch started at line 12. With context=2, it starts at line 13. The pattern is there and it makes sense.

 And 1:

 ```
 ➜  samplerepository git:(master) ✗ git diff -U1
diff --git a/README.md b/README.md
index 87f9baa..51c8a2d 100644
--- a/README.md
+++ b/README.md
@@ -14,4 +14,2 @@ Truces luctuque cognovit, cum lanam ordine vereri relinquunt sit munere quidam.
 Solent **torvi clamare successit** ille memores rogum; serpens egi caelo,
-moventem gelido volucrum reddidit fatalia, *in*. Abdit instant et, et hostis
-amores, nec pater formosus mortis capiat eripui: ferarum extemplo. Inmeritas
 favilla qui Dauno portis Aello! Fluit inde magis vinci hastam amore, mihi fama
 ```

 All is well. All matches. The starting line always match the header numbers. The pattern continues

 Now, with 0:

```
➜  samplerepository git:(master) ✗ git diff -U0
diff --git a/README.md b/README.md
index 87f9baa..51c8a2d 100644
--- a/README.md
+++ b/README.md
@@ -15,2 +14,0 @@ Solent **torvi clamare successit** ille memores rogum; serpens egi caelo,
-moventem gelido volucrum reddidit fatalia, *in*. Abdit instant et, et hostis
-amores, nec pater formosus mortis capiat eripui: ferarum extemplo. Inmeritas
```

Now, we got a (-15,2 +14,0). The first 15 makes sense as it corresponds to the removed "moventem", but there is a 14 right besides it. Which is no way what we expected it to be. From what we learned before we expected two 15s here.

And this is the problem with our assumption (and our implementation). It looks like (from doing other experiments locally and on git providers) that when the patch_size (either head or base) is ZERO, then the patch_start will be the previous number.

If we think about it, it is not exactly wrong: they said there were no lines from head in the patch. So they are not lying. I guess this falls under undefined behavior.

It is a bit confusing, though. I would expected a 15, 0 to be there instead.

So, basically, every time there is a zero-size context (on either base or head), we are doing the wrong thing and thinking the corresponding head line is 14 (in the example above). This breaks our python implementation

This is not a local thing only. Bitbucket, Github and Gitlab all follow this same flow (which makes sense as this is part of the unified format).

Luckily we only get this problem on zero-size contexts. And only if the particular diff segment has either only removals or only additions (the "problem" goes away if there is both, which happens when a line is changed)

The default context size is 3, so we never get this problem by default. The only time this problem happens is when we force context to be zero. Which is what we do on Bitbucket.

So we have two possibilities: we fix out implementation (which we should do either way), or we stop forcing contexts to be zero size. Stopping is the quicker fix for now, and doesn't risk breaking Github stuff in order to get Bitbucket stuff working (which is a relatively bad deal).

So we are just changing our context to be 1