Removing bad "token fallback" behaviors
On the worker, we have this behavior:

```
    {
        TokenType.read: admin_bot or get_config(repo.service, "bots", "read"),
        TokenType.admin: admin_bot,
        TokenType.comment: get_config(repo.service, "bots", "comment"),
        TokenType.status: admin_bot or get_config(repo.service, "bots", "status"),
    }
```

where we use different bots for different situations

Notice a few of them have a "fallback". If the `admin_bot` is None, we use a public bot on that case (for public repos only)

Even after that fallback, there is a second fallback:

```
    @property
    def token(self):
        if not self._token:
            self._token = self._oauth_consumer_token()
        return self._token
```

This fallback uses oauth key and secrets as token. It is a remnant of legacy where this was used for some reason.

But on new app, none of the actually made calls need to use those as tokens (except for `get_authenticated_user`, but that one already uses the oauth credentials directly)

So what we want to make is so that, if the "first" fallbacks don't do enough, the "token property" fallback does not try to do anything and we instead raise a `TorngitError` if in the end there are simply no tokens to make the call.

So we are making the `def api` call check for "there is at least a valid token otherwise raise", and another method can accept calls with or without tokens (so the `get_authenticated_user` can call that method).

When we do this, we expect to stop more calls before they are done, avoiding problems with Github.